Update WP Browser.jsx
added the port to use for the socket connection as a variable. needed this to do, to have a workaround for apache virtual host issues: if the virtual host is not the default configured host for the server, the connection would fail. a workaround is to let apache listen on a special port that only the desired virtual host is listening to.